Restructing of code
