+++ src/app/clubs/page.tsx
@@ -1,15 +1,17 @@
+/**
+ * Club Committe Page
+ *
+ * fetches data from /json/general/clubs
+ * Utilizes ClubCard Component
+ * Utilizes CircularProgress Component
+ */
+
 "use client";
 import ClubCard from "@/components/ClubCard/ClubCard";
 import React, { useEffect, useState } from "react";
 import styles from "./clubs.module.css";
 import { CircularProgress } from "@mui/material";
-
-interface Club {
-  name: string;
-  motto: string;
-  facultyIncharge: string;
-  coordinator: { name: string }[];
-}
+import { Club } from "@/types/common.types";
 
 const Clubs: React.FC = () => {
   const [clubs, setClubsData] = useState<Club[] | null>(null);
