+++ src/app/cmp/page.tsx
@@ -1,3 +1,11 @@
+/**
+ * CMPL Page
+ *
+ * Uses LabPage Component
+ *
+ */
+
+
 "use client";
 
 import LabPage from "@/components/LabComponent/LabComponent";
