+++ src/app/events/page.tsx
@@ -1,28 +1,29 @@
-"use client";
+/**
+ * Events Page
+ *
+ * fetches dara from
+ * utilizes NoticeSection
+ *
+ *
+ *
+ */
 
-import React, { useEffect, useState } from "react";
+
+"use client";;
+import { useEffect, useState } from "react";
 import Grid from "@mui/material/Grid2";
 import Box from "@mui/material/Box";
 import { Typography } from "@mui/material";
 import styles from "../notices/notices.module.css";
-import nextConfig from "../../../next.config";
-import { validURL } from "../../types/validator";
 import NoticeSection from "@/components/NoticeSection/NoticeSection";
-interface Item {
-  title: string;
-  link: string;
-  date?: string;
-  isNew?: boolean;
-  text?: string;
-}
-
+import { EventItem } from "@/types/common.types";
 
 
 const News = () => {
   const [loading, setLoading] = useState<boolean>(true);
   const [error, setError] = useState<string | null>(null);
-  const [oldNotices, setOldNotices] = useState<Item[]>([]);
-  const [newNotices, setNewNotices] = useState<Item[]>([]);
+  const [oldNotices, setOldNotices] = useState<EventItem[]>([]);
+  const [newNotices, setNewNotices] = useState<EventItem[]>([]);
 
   useEffect(() => {
     document.title = "Events | IIIT Tiruchirappalli ";
@@ -33,7 +34,7 @@ const News = () => {
 
         const data = await response.json();
 
-        const d: Item[] = data.data;
+        const d: EventItem[] = data.data;
         const latest = d
           .filter((x) => x.isNew)
           .sort((a, b) => new Date(b.date || "").getTime() - new Date(a.date || "").getTime());
