+++ src/types/achievements.types.ts
@@ -1,0 +1,8 @@
+
+export interface Item {
+  title: string;
+  link: string;
+  date?: string;
+  isNew?: boolean;
+  text?: string;
+}
