+++ src/types/nav.types.ts
@@ -11,4 +11,3 @@ export interface SubMenuItem {
     submenu?: SubMenuItem[];
     icon?: string;
   }
-
